Re-refactored Unit Test classes
