+++ ExpenseTracker/ExpenseTracker.Business.Tests/Base/BaseBudgetTest.cs
@@ -2,9 +2,9 @@
 using Dbo = ExpenseTracker.Persistence.Context.DbModels;
 using System.Collections.Generic;
 
-namespace ExpenseTracker.Business.Tests
+namespace ExpenseTracker.Business.Tests.Base
 {
-    public class BaseBudgetTest : BaseQueryTest
+    public class BaseBudgetTest : BaseTest
     {
         protected int DefaultTestBudgetId = -1;
